+++ lib/src/exports/helper_methods/bottom_sheet_border_rad.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-BorderRadius curveTopRightLeftCornerBy(double rad) {
+BorderRadius curveTopRightLeftCornerByRad({double rad = 20}) {
   return BorderRadius.only(
     topLeft: Radius.circular(rad),
     topRight: Radius.circular(rad),